pyproject.toml
[project]
name = "research-workflow"
version = "0.1.0"
description = "AI-powered research paper management for academic writing"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Jonas Tischer", email = "jonas.tischer@googlemail.com" }
]
keywords = ["research", "thesis", "citations", "ai", "claude", "academic"]

dependencies = [
    "marker-pdf>=1.0.0",
    "pymupdf>=1.24.0",
    "watchdog>=4.0.0",
    "google-genai>=1.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]
# For automated summaries (optional - Claude Code can do this directly)
auto-summarize = ["anthropic>=0.40.0"]

[project.scripts]
research-search = "src.search:main"
research-watch = "src.watcher:main"

[project.urls]
Repository = "https://github.com/JonasTischer/research-workflow"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
